refactor(drizzle): split order persistence helpers
- move Drizzle order row decoding, filtering, hydration, and item writes into a repository-local persistence helper
- keep DrizzleOrderRepository focused on port wiring and persistence error translation
- reduce the Fallow order repository hotspot while preserving the existing adapter behavior
